+++ packages/app/src/server/routes/apiv3/index.js
@@ -1,6 +1,4 @@
 import loggerFactory from '~/utils/logger';
-import * as userActivation from './user-activation';
-import injectUserRegistrationOrderByTokenMiddleware from '../../middlewares/inject-user-registration-order-by-token-middleware';
 
 const logger = loggerFactory('growi:routes:apiv3'); // eslint-disable-line no-unused-vars
 
@@ -54,12 +52,6 @@ module.exports = (crowi) => {
   router.use('/staffs', require('./staffs')(crowi));
 
   router.use('/forgot-password', require('./forgot-password')(crowi));
-
-  router.post('/complete-registration',
-    injectUserRegistrationOrderByTokenMiddleware,
-    userActivation.completeRegistrationRules(),
-    userActivation.validateCompleteRegistration,
-    userActivation.completeRegistrationAction(crowi));
 
   return router;
 };
