+++ packages/app/src/server/routes/index.js
@@ -199,6 +199,7 @@ module.exports = function(crowi, app) {
   app.use('/user-activation', express.Router()
     .get('/:token', apiLimiter, applicationInstalled, injectUserRegistrationOrderByTokenMiddleware, userActivation.form)
     .use(userActivation.handleHttpErrosMiddleware));
+  app.post('/user-activation/complete-registartion', apiLimiter, applicationInstalled, injectUserRegistrationOrderByTokenMiddleware, csrf, userActivation.completeRegistrationRules(), userActivation.validateCompleteRegistrationForm, userActivation.completeRegistrationAction(crowi));
   app.post('/user-activation/register', apiLimiter, applicationInstalled, csrf, userActivation.registerRules(), userActivation.validateRegisterForm, userActivation.registerAction(crowi));
 
   app.get('/share/:linkId', page.showSharedPage);
